Update to accommodate files grouped togeather
Based on the communities feedback we want to have the template, js, and
styles grouped togeather in a single folder.  The only option that makes
since to keep this feature is to have repetition in files and folders on
disk. While verbose this solution achieves the findability issue and keeps
the spirit of the original "pods".