Deprecation RFC Template
This is a specialized version of the RFC template, intended to
be used when a proposal suggests that a feature be
deprecated and replaced. It is more focused on the impact
of removals, but still fleshed out the details of the proposal
enough to get feedback and input from people who would
care about the change.