Clarify namespacing for addons invoking their own components.
The Core Team decided that it would be most clear if addons invoked 
helpers and components with the same namespacing to be used by
consumers of the addon. Without a namespace, components could be 
single non-dasherized words, which is verboten. Furthermore, this
decision makes templates more portable - it will be easier for an 
addon to adopt a component from a consumer, for instance.