Clarify the summary slightly
Previous wording here could be (mis)read to take this as implying that
our support is subpar or "second-class" despite its being hgih-quality
and comparable to the state of much of the rest of the front end
ecosystem (ahead in some ways, behind in others). This small revision
clarifies that, by emphasizing the pieces we *want* to emphasize: making
support an official policy, and bringing it within the Framework core
umbrella.