Clarify deprecation and interface in 0649
The previous version of the text suggested that the new interface with
`for` and `since` would be mandatory *immediately*. That, however, would
have been a breaking change which itself did not follow Ember's SemVer
and deprecation rules! Clarify that invoking `deprecate` without the
`for` or `since` options is deprecated until 4.0, and show this clearly
in the provided TypeScript interfaces.