Amend RFC 0821 given implementation tweaks
Authoring the types for DefinitelyTyped revealed a number of issues,
which were also surfaced previously while working on Ember's internal
versions of these types. Accordingly, align the types as defined here
with Ember's internal representation of them. Also add a `FullName`
export for convenience.

As a bonus, add a table of contents and make sure the order of the
sections matches the order described in prose.